backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime>=1.17.1
opencv-python-headless>=4.8.1.78
numpy>=1.26.4
supabase==2.0.3
python-dotenv==1.0.0
